fix: Address residual rounding issues in msats conversions and ledger entries
Code Analysis: Float vs. Decimal and Fractional Msats in Ledger Entries
TL;DR
The 188 msats residual is not caused by Python float precision loss. It is caused by non-integer Decimal msats values (produced by HIVE_amount × rate × 1000) being stored unrounded in ledger entries, while the corresponding Lightning payments use integer msats. The float usages present are largely safe, but several are architectural smells.

+++ src/v4vapp_backend_v2/models/payment_models.py
@@ -177,12 +177,12 @@ class Payment(TrackedBaseModel):
         if self.fee_msat:
             self.fee_conv = CryptoConversion(
                 conv_from=Currency.MSATS,
-                value=float(self.fee_msat),
+                value=int(self.fee_msat),
                 quote=quote,
             ).conversion
         self.conv = CryptoConversion(
             conv_from=Currency.MSATS,
-            value=float(self.value_msat + self.fee_msat),
+            value=int(self.value_msat + self.fee_msat),
             quote=quote,
         ).conversion
 
@@ -302,16 +302,14 @@ class Payment(TrackedBaseModel):
             return "Unknown"
 
         route_fees_ppm = self.route_fees_ppm
-        ans = " -> ".join(
-            [
-                (
-                    f"{hop.alias}"
-                    if route_fees_ppm.get(hop.pub_key) is None
-                    else f"{hop.alias} ({route_fees_ppm.get(hop.pub_key):.0f} ppm)"
-                )
-                for hop in self.route
-            ]
-        )
+        ans = " -> ".join([
+            (
+                f"{hop.alias}"
+                if route_fees_ppm.get(hop.pub_key) is None
+                else f"{hop.alias} ({route_fees_ppm.get(hop.pub_key):.0f} ppm)"
+            )
+            for hop in self.route
+        ])
         return ans
 
     # Methods from Payment
